pr feedback and add unit tests
